secondary_server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv

Werkzeug==2.0.1

